+++ Entities/AHuman.cpp
@@ -3075,8 +3075,12 @@ void AHuman::UpdateAI()
 int AHuman::OnPieMenu(Actor *pieMenuActor) {
 	int status = Actor::OnPieMenu(pieMenuActor);
 
-    if (status >= 0 && m_pFGArm && m_pFGArm->IsAttached() && m_pFGArm->HoldsDevice()) {
-        return m_pFGArm->GetHeldDevice()->OnPieMenu(pieMenuActor);
+    // Note: This is a bit ugly, but it should make this function output different error statuses based on whether the AHuman's OnPieMenuFunction fails, or its weapons' do, though the specifics can't be sussed out by the error alone.
+    if (m_pFGArm && m_pFGArm->IsAttached() && m_pFGArm->HoldsDevice()) {
+        status += m_pFGArm->GetHeldDevice()->OnPieMenu(pieMenuActor);
+    }
+    if (m_pBGArm && m_pBGArm->IsAttached() && m_pBGArm->HoldsDevice()) {
+        status += m_pBGArm->GetHeldDevice()->OnPieMenu(pieMenuActor);
     }
 
 	return status;